Fix for issue 20823 (#44)
* Fix for issue 20823
1. When Configure: false, and no HostedClusterspec, make sure to log an error, set a condition and requeue gracefully
2. On Delete, make sure that the finalizer does not get stuck, waiting for infrastructure to delete that should not be present. Also make sure the manifestWork delete continues when there is NO manifestWork.
3. Add test case to cover this scenario. test includes a check for the NEW condition and that delete is successful.


* catch error message from condition update

+++ pkg/controllers/hypershiftdeployment_controller.go
@@ -444,7 +444,8 @@ func (r *HypershiftDeploymentReconciler) destroyHypershift(hyd *hypdeployment.Hy
 		}
 	}
 
-	if hyd.Spec.Override != hypdeployment.InfraOverrideDestroy {
+	if hyd.Spec.Override != hypdeployment.InfraOverrideDestroy &&
+		hyd.Spec.Infrastructure.Configure {
 		// Infrastructure is the last step
 		if hyd.Spec.Infrastructure.Platform.AWS != nil {
 			if result, err := r.destroyAWSInfrastructure(hyd, providerSecret); err != nil {
